Uses spawn instead of exec and fixes normaliztion
The switch to spawn is merely a means of being more explicit.

Normalization was breaking when there was only General
Info returned (a GZip in this case).  Now it checks
to make sure the "track" array that's being normalized
isn't just a single track that can be parsed on its
own.